requirements.txt
bcc==0.18.0
blinker==1.4
Brlapi==0.8.3
certifi==2024.12.14
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
greenlet==3.1.1
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
playwright==1.49.0
pop-transition==1.1.2
praw==7.8.1
prawcore==2.4.0
protobuf==3.12.4
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydbus==0.6.0
pyee==12.0.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
repolib==2.2.1
repoman==1.4.0
requests==2.32.3
schedule==1.2.2
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
SQLAlchemy==2.0.36
systemd-python==234
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
update-checker==0.18.0
urllib3==2.2.3
wadllib==1.3.6
websocket-client==1.8.0
xdg==5
xkit==0.0.0
zipp==1.0.0
